Mysql-shard chore: Move those to new package
